refactor: optimize domain discovery and upload handling
- Updated the DeviceOverview component to enhance file size distribution and top domains visualization.
- Optimized SQL queries in the domain discovery and device overview routes to improve performance.
- Refactored the upload route to streamline the upload process, enhancing error handling and cleanup procedures.
